Implement matrix configuration and modify factory beans
Refactored the code to implement matrix configuration and modifications to various factory beans in spring configuration classes. The matrix configuration is added to allow dynamic creation of different types of matrices. The existing factory beans were modified to make them compatible with the new matrix configuration. Removed the unnecessary instance variables from the factory beans and introduced the '@Lazy' annotation for the bean methods to delay their instantiation until they are requested.