Add option to define number of threads for simulation
Implemented a new parameter, THREADS_COUNT, allowing the user to define the number of threads to be used for the simulation. This gives the user control over resource allocation. If THREADS_COUNT is not defined, the number of available processors will be used as before. Changes also include modifying the car index retrieval method in TPS_WritableTrip class.